+++ example/src/config/GlobalConst.js
@@ -1,0 +1,14 @@
+/*
+ * @Description: GlobalConst
+ * @Author: wanglh
+ * @LastEditors: wanglh
+ * @Date: 2019-04-23 17:43:33
+ * @LastEditTime: 2019-04-23 17:57:11
+ */
+
+// 字符串常量
+const GlobalString = {
+    DemoEmptyString: '这家伙太懒了，还没有添加demo哦~',
+};
+
+export default GlobalString;